split out tar into its own module
